Move test to trainer restore suite + test eventual success vs failure cases
