+++ python/ray/train/tests/test_tune.py
@@ -191,36 +191,6 @@ def test_reuse_checkpoint(ray_start_4_cpus):
     assert len(trial_dfs[0]["training_iteration"]) == 5
 
 
-def test_retry_with_max_failures(ray_start_4_cpus):
-    def train_func():
-        ckpt = session.get_checkpoint()
-        restored = bool(ckpt)  # Does a previous checkpoint exist?
-        itr = 0
-        if ckpt:
-            ckpt = ckpt.to_dict()
-            itr = ckpt["iter"] + 1
-
-        for i in range(itr, 4):
-            if i == 2 and not restored:
-                raise RuntimeError("try to fail me")
-            session.report(
-                dict(test=i, training_iteration=i),
-                checkpoint=Checkpoint.from_dict(dict(iter=i)),
-            )
-
-    trainer = DataParallelTrainer(
-        train_func,
-        backend_config=TestConfig(),
-        scaling_config=ScalingConfig(num_workers=1),
-        run_config=RunConfig(failure_config=FailureConfig(max_failures=-1)),
-    )
-
-    result = trainer.fit()
-    assert not result.error
-    checkpoint = result.checkpoint.to_dict()
-    assert checkpoint["iter"] == 3
-
-
 def test_restore_with_new_trainer(ray_start_4_cpus, tmpdir, propagate_logs, caplog):
     def train_func(config):
         raise RuntimeError("failing!")
